# backend/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nutrition Backend - Auth, Profile, Target & Meal API
  version: 0.1.0
servers:
  - url: /api/v1
...
paths:
...
  /targets:
    get:
      tags: [Target]
      summary: List current user's targets
      description: >
        List all nutrition targets of the currently authenticated user
        (optionally paginated with limit/offset). The newest targets
        (by created_at) are returned first. Each target represents a
        daily set of 17 nutrient goals.
      security:
        - AccessTokenCookieAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
          required: false
          description: Maximum number of targets to return (default is implementation-defined)
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: Number of items to skip before starting to collect the result set
      responses:
        '200':
          description: List of user's targets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetListResponse'
        '401':
          description: Unauthorized (missing or invalid access token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Target]
      summary: Create a new nutrition target
      description: >
        Create a new daily nutrition target definition for the current user.

        The server uses the user's Profile (sex, birthdate, height, weight) plus
        the provided goal information (goal_type, activity_level, etc.) to generate
        a set of recommended daily intakes for 17 nutrients using a model
        (e.g., LLM or rules). If this is the first target for the user,
        it becomes the active target.
      security:
        - AccessTokenCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTargetRequest'
      responses:
        '201':
          description: Target created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '400':
          description: Invalid input (e.g., missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (missing or invalid access token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Maximum number of targets reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /targets/active:
    get:
      tags: [Target]
      summary: Get active target
      description: >
        Return the active nutrition target for the currently authenticated user.
        If no target is active or if no targets exist, an error is returned.
      security:
        - AccessTokenCookieAuth: []
      responses:
        '200':
          description: Active target of the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '401':
          description: Unauthorized (missing or invalid access token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active target or no targets exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /targets/{target_id}:
    get:
      tags: [Target]
      summary: Get target by ID
      description: >
        Get a specific nutrition target of the current user by its ID.
      security:
        - AccessTokenCookieAuth: []
      parameters:
        - in: path
          name: target_id
          required: true
          description: Target ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Target detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '401':
          description: Unauthorized (missing or invalid access token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Target not found or does not belong to the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [Target]
      summary: Partially update a target
      description: >
        Partially update fields of an existing nutrition target.
        Only fields present in the request body are updated.
        When nutrient values are updated, their source is changed to "manual".
      security:
        - AccessTokenCookieAuth: []
      parameters:
        - in: path
          name: target_id
          required: true
          description: Target ID (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTargetRequest'
      responses:
        '200':
          description: Updated target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '400':
          description: Invalid nutrient code or invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (missing or invalid access token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Target not found or does not belong to the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /targets/{target_id}/activate:
    post:
      tags: [Target]
      summary: Activate a target
      description: >
        Activate the specified target for the current user.
        Any other targets of the same user will be set to inactive.
      security:
        - AccessTokenCookieAuth: []
      parameters:
        - in: path
          name: target_id
          required: true
          description: Target ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Activated target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetResponse'
        '401':
          description: Unauthorized (missing or invalid access token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Target not found or does not belong to the current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========
  # Meal
  # ==========
...
components:
  securitySchemes:
    AccessTokenCookieAuth:
      type: apiKey
      in: cookie
      name: ACCESS_TOKEN
      description: HttpOnly access token cookie

    RefreshTokenCookieAuth:
      type: apiKey
      in: cookie
      name: REFRESH_TOKEN
      description: HttpOnly refresh token cookie
  schemas:
...
    GoalType:
      type: string
      description: Target goal type.
      enum:
        - weight_loss
        - maintain
        - weight_gain
        - health_improve
    ActivityLevel:
      type: string
      description: User's activity level used for target calculation.
      enum:
        - low
        - normal
        - high
    NutrientCode:
      type: string
      description: >
        Code for one of the 17 nutrients tracked in daily targets.
      enum:
        # Energy sources
        - carbohydrate
        - fat
        - protein
        # Vitamins
        - vitamin_a
        - vitamin_b_complex
        - vitamin_c
        - vitamin_d
        - vitamin_e
        - vitamin_k
        # Minerals
        - calcium
        - iron
        - magnesium
        - zinc
        - sodium
        - potassium
        # Others
        - fiber
        - water
    NutrientSource:
      type: string
      description: Source of the target nutrient value.
      enum:
        - llm
        - manual
        - user_input
    TargetNutrient:
      type: object
      description: Target value for a single nutrient.
      properties:
        code:
          $ref: '#/components/schemas/NutrientCode'
        amount:
          type: number
          format: float
          description: Target amount of the nutrient.
          example: 150.0
        unit:
          type: string
          description: Unit of the amount (e.g. "g", "mg", "kcal", "ml").
          example: g
        source:
          $ref: '#/components/schemas/NutrientSource'
      required:
        - code
        - amount
        - unit
        - source
    TargetResponse:
      type: object
      description: A single nutrition target definition.
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        goal_type:
          $ref: '#/components/schemas/GoalType'
        goal_description:
          type: string
          nullable: true
        activity_level:
          $ref: '#/components/schemas/ActivityLevel'
        is_active:
          type: boolean
        nutrients:
          type: array
          items:
            $ref: '#/components/schemas/TargetNutrient'
        llm_rationale:
          type: string
          nullable: true
        disclaimer:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - title
        - goal_type
        - activity_level
        - is_active
        - nutrients
        - created_at
        - updated_at
    TargetListResponse:
      type: object
      description: List of targets for the current user.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TargetResponse'
      required:
        - items
    CreateTargetRequest:
      type: object
      description: Request body to create a new target.
      properties:
        title:
          type: string
          description: Human-friendly name of the target.
          example: 'Lean bulk (spring 2025)'
        goal_type:
          $ref: '#/components/schemas/GoalType'
        goal_description:
          type: string
          nullable: true
          description: Optional free-form description of the goal.
        activity_level:
          $ref: '#/components/schemas/ActivityLevel'
      required:
        - title
        - goal_type
        - activity_level
    UpdateTargetNutrient:
      type: object
      description: Partial update for a single nutrient in the target.
      properties:
        code:
          $ref: '#/components/schemas/NutrientCode'
        amount:
          type: number
          format: float
          nullable: true
          description: New amount. If omitted or null, the amount is not changed.
        unit:
          type: string
          nullable: true
          description: New unit. If omitted or null, the unit is not changed.
      required:
        - code
    UpdateTargetRequest:
      type: object
      description: Partial update for a target definition.
      properties:
        title:
          type: string
          nullable: true
        goal_type:
          $ref: '#/components/schemas/GoalType'
        goal_description:
          type: string
          nullable: true
        activity_level:
          $ref: '#/components/schemas/ActivityLevel'
        llm_rationale:
          type: string
          nullable: true
        disclaimer:
          type: string
          nullable: true
        nutrients:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/UpdateTargetNutrient'
      additionalProperties: false

    # ==========
    # Meal
    # ==========
...
    ErrorResponse:
      type: object
      description: Generic error response.
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - EMAIL_ALREADY_IN_USE
                - INVALID_CREDENTIALS
                - UNAUTHORIZED
                - USER_NOT_FOUND
                - PROFILE_NOT_FOUND
                - TARGET_NOT_FOUND
                - MAX_TARGETS_REACHED
                - VALIDATION_ERROR
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
      required:
        - error